Create gtn.py
When your program starts, the game immediately begins in the range [0, 100). When the game ends (because the player either wins or runs out of guesses), a new game with the same range as the last one immediately begin by calling new_game(). Whenever the player clicks one of the range buttons, the current game should stop and a new game with the selected range should begin.